fix: extract util
